move files for Explorer base
