+++ src/components/Explorer/Base/QueryBuilder/FilterSlider.jsx
