+++ src/components/Explorer/Base/QueryBuilder/d3RangeSlider.js
