+++ src/components/Explorer/Base/Result/Chart/ChartHelpers.jsx
