+++ src/components/Explorer/Base/Result/Chart/GenericChart.jsx
