+++ src/components/Explorer/Base/Result/ResultPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ExternalLink } from "../../../common/Helpers";
+import { ExternalLink } from "../../../../common/Helpers";
 import GenericChart, {
     renderChart as renderGenericChart
 } from './Chart/GenericChart';
